requirements.txt
asgiref==3.8.0
crispy-bootstrap5==2024.2
Django==5.0.3
django-crispy-forms==2.1
django-multiselectfield==0.1.12
numpy==1.26.4
OpenTimelineIO==0.16.0
pandas==2.2.1
pyaaf2==1.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tzdata==2024.1
Pillow
